requirements.txt
alembic==1.7.6
aniso8601==8.1.0
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.3
apache-airflow-providers-imap==2.2.0
apache-airflow-providers-postgres==4.0.1
apache-airflow-providers-sqlite==2.1.0
apispec==3.3.2
app==0.0.1
appdirs==1.4.4
appnope==0.1.2
argcomplete==2.0.0
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
blinker==1.4
blis==0.4.1
cachelib==0.6.0
catalogue==1.0.0
cattrs==1.10.0
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
cmake==3.18.4.post1
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
connexion==2.12.0
croniter==1.3.4
cryptography==36.0.1
cycler==0.10.0
cymem==2.0.4
dbconfig==0.3.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
Django==3.1.4
dlib==19.21.1
dnspython==2.2.0
docutils==0.16
elasticsearch==7.13.4
elasticsearch-dbapi==0.2.9
elasticsearch-dsl==7.4.0
email-validator==1.1.3
enum34==1.1.10
Faker==13.3.3
Flask==1.1.2
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-MySQL==1.5.1
Flask-OpenID==1.3.0
Flask-RESTful==0.3.8
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.19.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httpcore==0.14.7
httpx==0.22.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.11.2
inflection==0.5.1
ipython==7.27.0
iso8601==1.0.2
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==1.1.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mccabe==0.6.1
murmurhash==1.0.4
networkx==2.6.2
numpy==1.19.4
opencv-python==4.5.3.56
packaging==21.3
pandas==1.4.1
parso==0.8.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
plaidml==0.7.0
postgres==4.0
preshed==3.0.4
prison==0.2.1
prompt-toolkit==3.0.20
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytools==2021.1
pytz==2020.4
pytzdata==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rich==11.2.0
scikit-image==0.18.2
scipy==1.6.1
setproctitle==1.2.2
six==1.15.0
smmap==5.0.0
sniffio==1.2.0
spacy==2.3.2
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.1
srsly==1.0.4
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
thinc==7.4.1
tifffile==2021.8.8
toml==0.10.2
tqdm==4.53.0
traitlets==5.1.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.2
vaderSentiment==3.3.2
wasabi==0.8.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.7.0
